Expose keepOpenOnPeerClosed for outgoing TCP streams (#22163)
I'm not too familiar with this part of the codebase but it seems to make sense
to expose this option. Not sure what the most sensible default is: accidentally
enabling it might cause missing data when the server is eager in closing the
connections, accidentally disabling it might cause the outgoing side of the
connection to 'leak'.